Refs #20739: Make monitor service interfaces private (except IProxyQueryable)
